+++ lib/mangoapps/modules/recognitions/get_award_feeds.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module MangoApps
+  class Client
+    module Recognitions
+      module GetAwardFeeds
+        def get_award_feeds(params = {})
+          get("v2/recognitions/get_award_feeds.json", params: params)
+        end
+      end
+    end
+  end
+end