+++ lib/mangoapps/modules/wikis/get_wikis.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module MangoApps
+  class Client
+    module Wikis
+      module GetWikis
+        def get_wikis(params = {})
+          get("v2/wikis.json", params: params)
+        end
+      end
+    end
+  end
+end